+++ picasso/gaussmle.py
@@ -865,7 +865,7 @@ def locs_from_fits(
             "iterations": iterations.astype(np.uint32),
         }
     )
-    if hasattr(identifications, "n_id"):
+    if "n_id" in identifications.columns:
         locs["n_id"] = identifications.n_id.astype(np.uint32)
         locs.sort_values(by=["n_id"], kind="quicksort", inplace=True)
     else: