+++ picasso/gui/render.py
@@ -769,7 +769,7 @@ class DatasetDialog(QtWidgets.QDialog):
 
             # adjust group color if needed
             if len(self.window.view.locs) == 1:
-                if hasattr(self.window.view.locs[0], "group"):
+                if "group" in self.window.view.locs[0].columns:
                     self.window.view.group_color = (
                         self.window.view.get_group_color(
                             self.window.view.locs[0]
@@ -2174,7 +2174,7 @@ class G5MDialog(QtWidgets.QDialog):
         self.window = window
         self.nena = None
         self.channel = channel
-        self.flag_3D = hasattr(self.window.view.locs[0], "z")
+        self.flag_3D = "z" in self.window.view.locs[0].columns
         self.setWindowTitle("Molecular mapping (G5M) parameters")
 
         vbox = QtWidgets.QVBoxLayout(self)
@@ -2376,7 +2376,7 @@ class G5MDialog(QtWidgets.QDialog):
             return
         try:
             cluster_sizes = pd.read_csv(path)
-            if hasattr(self.window.view.locs[self.channel], "z"):  # 3D
+            if "z" in self.window.view.locs[self.channel].columns:  # 3D
                 column = "Volume (LP^3)"
                 thresh = 12 * 4 / 3 * np.pi * 2.98**2 * (2.98 * 2.5)
             else:  # 2D
@@ -2562,14 +2562,14 @@ class TestClustererDialog(QtWidgets.QDialog):
         self.view.update_scene()
 
     def on_xz_proj(self) -> None:
-        if self.view.locs is not None and hasattr(self.view.locs, "z"):
+        if self.view.locs is not None and "z" in self.view.locs.columns:
             self.view.ang = [1.5708, 0, 0]  # 90 deg rotation
         else:
             self.view.ang = None
         self.view.update_scene()
 
     def on_yz_proj(self) -> None:
-        if self.view.locs is not None and hasattr(self.view.locs, "z"):
+        if self.view.locs is not None and "z" in self.view.locs.columns:
             self.view.ang = [0, 1.5708, 0]  # 90 deg rotation
         else:
             self.view.ang = None
@@ -2606,7 +2606,7 @@ class TestClustererDialog(QtWidgets.QDialog):
             self.view.group_color = self.window.view.get_group_color(locs)
 
         # scale z axis if applicable
-        if hasattr(locs, "z"):
+        if "z" in locs.columns:
             locs.z /= pixelsize
         return locs
 
@@ -2953,7 +2953,7 @@ class TestClustererView(QtWidgets.QLabel):
         ):  # two channels, all locs and clustered locs
             channel = self.dialog.channel
             all_locs = self.dialog.window.view.picked_locs(channel)[0]
-            if hasattr(all_locs, "z"):
+            if "z" in all_locs.columns:
                 all_locs.z /= (
                     self.dialog.window.display_settings_dlg.pixelsize.value()
                 )
@@ -2975,7 +2975,7 @@ class TestClustererView(QtWidgets.QLabel):
         ):  # three channels, all locs, clustered locs and cluster centers
             channel = self.dialog.channel
             all_locs = self.dialog.window.view.picked_locs(channel)[0]
-            if hasattr(all_locs, "z"):
+            if "z" in all_locs.columns:
                 all_locs.z /= (
                     self.dialog.window.display_settings_dlg.pixelsize.value()
                 )
@@ -4611,7 +4611,7 @@ class RESIDialog(QtWidgets.QDialog):
         self.n_channels = len(self.locs)
         self.paths = window.view.locs_paths
         self.ndim = 2
-        if all([hasattr(_, "z") for _ in self.locs]):
+        if all(["z" in _.columns for _ in self.locs]):
             self.ndim = 3
 
         self.radius_xy = []
@@ -5473,8 +5473,9 @@ class FastRenderDialog(QtWidgets.QDialog):
                 factors.append(new_disp_nlocs / old_disp_nlocs)
             factor = np.mean(factors)  # to adjust contrast
         #  update view.group_color if needed:
-        if len(self.fractions) == 2 and hasattr(
-            self.window.view.locs[0], "group"
+        if (
+            len(self.fractions) == 2
+            and "group" in self.window.view.locs[0].columns
         ):
             self.window.view.group_color = self.window.view.get_group_color(
                 self.window.view.locs[0]
@@ -6000,7 +6001,7 @@ class View(QtWidgets.QLabel):
             self.median_lp = np.mean(
                 [np.median(locs.lpx), np.median(locs.lpy)]
             )
-            if hasattr(locs, "group"):
+            if "group" in locs.columns:
                 if len(self.group_color) == 0 and locs.group.size:
                     self.group_color = self.get_group_color(self.locs[0])
             disp_sett_dlg.parameter.clear()
@@ -6015,7 +6016,7 @@ class View(QtWidgets.QLabel):
             self.fit_in_view(autoscale=True)
             self.update_scene()
 
-        if hasattr(locs, "z"):
+        if "z" in locs.columns:
             # append z coordinates for slicing
             self.window.slicer_dialog.zcoord.append(locs.z)
             # unlock 3D settings
@@ -6281,7 +6282,7 @@ class View(QtWidgets.QLabel):
         self.all_locs[channel] = pd.concat(out_locs, ignore_index=True)
         self.locs[channel] = copy.copy(self.all_locs[channel])
 
-        if hasattr(self.all_locs[channel], "group"):
+        if "group" in self.all_locs[channel].columns:
             groups = np.unique(self.all_locs[channel].group)
             # In case a group is missing
             groups = np.arange(np.max(groups) + 1)
@@ -6297,7 +6298,7 @@ class View(QtWidgets.QLabel):
 
         See ``picasso.postprocess.link`` for more details."""
         channel = self.get_channel()
-        if hasattr(self.all_locs[channel], "len"):
+        if "len" in self.all_locs[channel].columns:
             QtWidgets.QMessageBox.information(
                 self, "Link", "Localizations are already linked. Aborting..."
             )
@@ -6315,7 +6316,7 @@ class View(QtWidgets.QLabel):
                     max_dark_time=max_dark,
                 )
                 status.close()
-                if hasattr(self.all_locs[channel], "group"):
+                if "group" in self.all_locs[channel].columns:
                     groups = np.unique(self.all_locs[channel].group)
                     groups = np.arange(np.max(groups) + 1)
                     np.random.shuffle(groups)
@@ -6393,7 +6394,7 @@ class View(QtWidgets.QLabel):
             "Applying DBSCAN. This may take a while.", self
         )
         # keep group info if already present
-        if hasattr(self.all_locs[channel], "group"):
+        if "group" in self.all_locs[channel].columns:
             locs = self.all_locs[channel].copy()
             locs["group_input"] = self.all_locs[channel].group
         else:
@@ -6520,7 +6521,7 @@ class View(QtWidgets.QLabel):
             "Applying HDBSCAN. This may take a while.", self
         )
         # keep group info if already present
-        if hasattr(self.all_locs[channel], "group"):
+        if "group" in self.all_locs[channel].columns:
             locs = self.all_locs[channel].copy()
             locs["group_input"] = self.all_locs[channel].group
         else:
@@ -6577,7 +6578,7 @@ class View(QtWidgets.QLabel):
 
         # get clustering parameters
         pixelsize = self.window.display_settings_dlg.pixelsize.value()
-        if any([hasattr(_, "z") for _ in self.all_locs]):
+        if any(["z" in _.columns for _ in self.all_locs]):
             flag_3D = True
         else:
             flag_3D = False
@@ -6655,7 +6656,7 @@ class View(QtWidgets.QLabel):
         status = lib.StatusDialog("Clustering localizations", self)
 
         # keep group info if already present
-        if hasattr(self.all_locs[channel], "group"):
+        if "group" in self.all_locs[channel].columns:
             locs = self.all_locs[channel].copy()
             locs["group_input"] = self.all_locs[channel].group
         else:
@@ -6683,7 +6684,7 @@ class View(QtWidgets.QLabel):
             "Performed basic frame analysis": frame_analysis,
             "Fraction of rejected locs (%)": rejected,
         }
-        if hasattr(self.all_locs[channel], "z"):
+        if "z" in self.all_locs[channel].columns:
             new_info["Clustering radius xy (nm)"] = radius_xy * pixelsize
             new_info["Clustering radius z (nm)"] = radius_z * pixelsize
         else:
@@ -6918,7 +6919,7 @@ class View(QtWidgets.QLabel):
         """Check whether the data has been grouped (clustered) in
         channel i."""
         locs = self.window.view.locs[channel]
-        if hasattr(locs, "group"):
+        if "group" in locs.columns:
             return True
         else:
             message = (
@@ -6987,7 +6988,7 @@ class View(QtWidgets.QLabel):
         locs = [self.picked_locs(_) for _ in range(n_channels)]
         dy = self.shifts_from_picked_coordinate(locs, "y")
         dx = self.shifts_from_picked_coordinate(locs, "x")
-        if all([hasattr(_[0], "z") for _ in locs]):
+        if all(["z" in _[0].columns for _ in locs]):
             dz = self.shifts_from_picked_coordinate(locs, "z")
         else:
             dz = None
@@ -8253,8 +8254,9 @@ class View(QtWidgets.QLabel):
             x2 = self.locs[channel2].x * pixelsize
             y1 = self.locs[channel1].y * pixelsize
             y2 = self.locs[channel2].y * pixelsize
-            if hasattr(self.locs[channel1], "z") and hasattr(
-                self.locs[channel2], "z"
+            if (
+                "z" in self.locs[channel1].columns
+                and "z" in self.locs[channel2].columns
             ):
                 z1 = self.locs[channel1].z
                 z2 = self.locs[channel2].z
@@ -8882,7 +8884,7 @@ class View(QtWidgets.QLabel):
                 if self._picks:
                     for i, pick in enumerate(self._picks):
                         # 3D
-                        if hasattr(all_picked_locs[0], "z"):
+                        if "z" in all_picked_locs[0].columns:
                             # k-means clustering
                             reply = ClsDlg3D.getParams(
                                 all_picked_locs,
@@ -8924,7 +8926,7 @@ class View(QtWidgets.QLabel):
                         reply = 3
                         while reply == 3:
                             # 3D
-                            if hasattr(all_picked_locs[0], "z"):
+                            if "z" in all_picked_locs[0].columns:
                                 # k-means clustering
                                 reply, nc, l_locs, c_locs = ClsDlg3D.getParams(
                                     all_picked_locs,
@@ -8993,7 +8995,7 @@ class View(QtWidgets.QLabel):
             dark = np.empty(len(clustered_locs))
 
             for i, pick_locs in enumerate(clustered_locs):
-                if not hasattr(pick_locs, "len"):
+                if "len" not in pick_locs.columns:
                     pick_locs = postprocess.link(
                         pick_locs,
                         self.infos[channel],
@@ -9595,7 +9597,7 @@ class View(QtWidgets.QLabel):
 
         # if slicing, show only current slice from every channel
         for i in range(len(locs)):
-            if hasattr(locs[i], "z"):
+            if "z" in locs[i].columns:
                 if self.window.slicer_dialog.slicer_radio_button.isChecked():
                     z_min = self.window.slicer_dialog.slicermin
                     z_max = self.window.slicer_dialog.slicermax
@@ -9697,13 +9699,13 @@ class View(QtWidgets.QLabel):
             )
 
         # if locs have group identity (e.g. clusters)
-        if hasattr(locs, "group") and locs.group.size:
+        if "group" in locs.columns and locs.group.size:
             locs = [locs[self.group_color == _] for _ in range(N_GROUP_COLORS)]
             return self.render_multi_channel(
                 kwargs, locs=locs, autoscale=autoscale, use_cache=use_cache
             )
         # if slicing, show only the current slice
-        if hasattr(locs, "z"):
+        if "z" in locs.columns:
             if self.window.slicer_dialog.slicer_radio_button.isChecked():
                 z_min = self.window.slicer_dialog.slicermin
                 z_max = self.window.slicer_dialog.slicermax
@@ -9972,7 +9974,7 @@ class View(QtWidgets.QLabel):
         # allow running even if no picks are present but group info is
         if len(self._picks) == 0:
             locs = self.all_locs[channel]
-            if not hasattr(locs, "group"):
+            if "group" not in locs.columns:
                 message = (
                     "No picks found. Please create picks or assign group "
                     "identity to localizations before calculating pick "
@@ -10003,7 +10005,7 @@ class View(QtWidgets.QLabel):
             progress.set_value(i + 1)
             if not len(pick_locs):
                 continue
-            if not hasattr(pick_locs, "len"):
+            if "len" not in pick_locs.columns:
                 pick_locs = postprocess.link(
                     pick_locs,
                     self.infos[channel],
@@ -10447,7 +10449,7 @@ class View(QtWidgets.QLabel):
                 )
             else:
                 self.plot_window = DriftPlotWindow(self)
-                if hasattr(self._drift[channel], "z"):
+                if "z" in self._drift[channel].columns:
                     self.plot_window.plot_3d(drift)
 
                 else:
@@ -10569,7 +10571,7 @@ class View(QtWidgets.QLabel):
             self.locs[channel]["x"] -= drift["x"].iloc[frames_].to_numpy()
             self.locs[channel]["y"] -= drift["y"].iloc[frames_].to_numpy()
             # If z coordinate exists, also apply drift there
-            if all([hasattr(_, "z") for _ in picked_locs]):
+            if all(["z" in _.columns for _ in picked_locs]):
                 self.all_locs[channel]["z"] -= (
                     drift["z"].iloc[frames].to_numpy()
                 )
@@ -10636,7 +10638,7 @@ class View(QtWidgets.QLabel):
         self.locs[channel]["x"] -= drift["x"].iloc[frames_].to_numpy()
         self.locs[channel]["y"] -= drift["y"].iloc[frames_].to_numpy()
 
-        if hasattr(drift, "z"):
+        if "z" in drift.columns:
             drift["z"] = -drift["z"]
             self.all_locs[channel]["z"] -= drift["z"].iloc[frames].to_numpy()
             self.locs[channel]["z"] -= drift["z"].iloc[frames_].to_numpy()
@@ -10665,8 +10667,8 @@ class View(QtWidgets.QLabel):
             self._drift[channel]["x"] += drift["x"]
             self._drift[channel]["y"] += drift["y"]
 
-            if hasattr(drift, "z"):
-                if hasattr(self._drift[channel], "z"):
+            if "z" in drift.columns:
+                if "z" in self._drift[channel].columns:
                     self._drift[channel]["z"] += drift["z"]
                 else:
                     self._drift[channel]["z"] = drift["z"]
@@ -10751,7 +10753,7 @@ class View(QtWidgets.QLabel):
 
         # automatically assign the group if circular picks are present
         if (
-            not hasattr(self.all_locs[0], "group")
+            "group" not in self.all_locs[0].columns
             and len(self._picks)
             and self._pick_shape == "Circle"
         ):
@@ -10759,7 +10761,7 @@ class View(QtWidgets.QLabel):
             locs = pd.concat(locs, ignore_index=True)
             self.all_locs[0] = locs
             remove_group = True
-        elif hasattr(self.all_locs[0], "group"):
+        elif "group" in self.all_locs[0].columns:
             remove_group = False
         else:
             QtWidgets.QMessageBox.information(
@@ -10930,7 +10932,7 @@ class View(QtWidgets.QLabel):
             rmsd = np.empty(n_picks)  # rmsd in each pick
             length = np.empty(n_picks)  # estimated mean bright time
             dark = np.empty(n_picks)  # estimated mean dark time
-            has_z = hasattr(picked_locs[0], "z")
+            has_z = "z" in picked_locs[0].columns
             if has_z:
                 rmsd_z = np.empty(n_picks)
             new_locs = []  # linked locs in each pick
@@ -10956,7 +10958,7 @@ class View(QtWidgets.QLabel):
                         rmsd_z[i] = np.sqrt(
                             np.mean((locs.z - np.mean(locs.z)) ** 2)
                         )
-                    if not hasattr(locs, "len"):
+                    if "len" not in locs.columns:
                         locs = postprocess.link(
                             locs, info, r_max=r_max, max_dark_time=t
                         )
@@ -11972,7 +11974,7 @@ class Window(QtWidgets.QMainWindow):
                         info,
                     )
 
-                    if hasattr(locs, "z"):
+                    if "z" in locs.columns:
                         z_min = locs["z"][in_view].min()
                         z_max = locs["z"][in_view].max()
                         z_mins.append(z_min)