+++ changelog.rst
@@ -8,6 +8,7 @@ Last change: 02-FEB-2026 CEST
 - Render: increase the speed of picking circular locs, picking similar and filter by number of localizations (numba implementation)
 - Reduced copying and conversion of DataFrames to numpy arrays (less memory usage)
 - Moved from merge sort to quick sort (often faster due to lower memory usage)
+- Use ``"col" in df.columns`` instead of ``hasattr(df, "col")`` to check for columns in DataFrames (better readability)
 
 0.9.4-5
 -------